tool/requirements.txt
docker-compose==1.6.2
pymongo==3.2.2
PyYAML==3.11
docker-py==1.8.1